Use stdout instead of stderr for general logging
Rollup writes additional information regarding file creation and
processing of files to `stderr`. However, these messages do not
indicate any actual errors. Instead, they are used as general
information. The original intent of these message was probably
stdout.

The only additional consideration is that Rollup itself can be used
when writing its output to `stdout`. Therefore, if we are using
stdout as the output mechanism, we write these general information
messages to `stderr`. In all other cases, we write to `stdout`.

We hit this issue downstream in DevTools, where Rollup was writing
to `stderr`, which consequently failed scripts that perform logic
along the lines of "if the script wrote to stderr it failed".